Removed the SSO authentication mechanism, and added the login page for authentication directly.
