+++ .idea/jarRepositories.xml
@@ -31,5 +31,10 @@
       <option name="name" value="maven3" />
       <option name="url" value="https://splunk.jfrog.io/splunk/ext-releases-local" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="MavenLocal" />
+      <option name="name" value="MavenLocal" />
+      <option name="url" value="file:/$MAVEN_REPOSITORY$/" />
+    </remote-repository>
   </component>
 </project>
