+++ src/main/resources/application.properties
@@ -4,8 +4,8 @@ server.servlet.context-path=/dashboard
 server.servlet.application-display-name=Dashboard
 
 #security.basic.enabled=false
-base.url=http://wlguatrosiapp01.natlib.govt.nz:${server.port}${server.servlet.context-path}
-#base.url=http://localhost:${server.port}${server.servlet.context-path}
+#base.url=http://wlguatrosiapp01.natlib.govt.nz:${server.port}${server.servlet.context-path}
+base.url=http://localhost:${server.port}${server.servlet.context-path}
 
 # Spring MVC configuration
 spring.mvc.view.prefix=/
