+++ src/main/webapp/app/login.html
@@ -1,21 +1,155 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta http-equiv="x-ua-compatible" content="ie=edge">
-  <title>Deposit Dashboard Login</title>
-  <link rel="icon" type="image/x-icon" href="favicon.ico" />
-</head>
 
-<body>
-  <script>
-    if (parent && parent.isParentAvailable && parent.isParentAvailable()) {
-      parent.loginCallback();
-    }else{
-      window.location.replace('${PATH_CONTEXT}');
-    }
-  </script>
-</body>
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="description" content="">
+  
+    <title>Signin Dashbard</title>
+    
+
+    <link rel="icon" type="image/x-icon" href="favicon.ico" />
+    <link rel="stylesheet" href="./dist/bootstrap.css">
+    <link rel="stylesheet" href="./dist/bootstrap-icons.css">
+    <link rel="stylesheet" href="./dist/icheck-bootstrap.min.css">
+    <link rel="stylesheet" href="./dist/toastr.min.css">
+    <meta name="theme-color" content="#7952b3">
 
 
+    <style>
+      .bd-placeholder-img {
+        font-size: 1.125rem;
+        text-anchor: middle;
+        -webkit-user-select: none;
+        -moz-user-select: none;
+        user-select: none;
+      }
+
+      @media (min-width: 768px) {
+        .bd-placeholder-img-lg {
+          font-size: 3.5rem;
+        }
+      }
+
+      html,
+      body {
+        height: 100%;
+      }
+
+      body {
+        display: flex;
+        align-items: center;
+        padding-top: 40px;
+        padding-bottom: 40px;
+        background-color: #f5f5f5;
+      }
+
+      .form-signin {
+        width: 100%;
+        max-width: 330px;
+        padding: 15px;
+        margin: auto;
+      }
+
+      .form-signin .checkbox {
+        font-weight: 400;
+      }
+
+      .form-signin .form-floating:focus-within {
+        z-index: 2;
+      }
+
+      .form-signin input[type="email"] {
+        margin-bottom: -1px;
+        border-bottom-right-radius: 0;
+        border-bottom-left-radius: 0;
+      }
+
+      .form-signin input[type="password"] {
+        margin-bottom: 10px;
+        border-top-left-radius: 0;
+        border-top-right-radius: 0;
+      }
+
+    </style>
+  </head>
+  <body class="text-center">
+    <main class="form-signin">
+      <div>
+        <img class="mb-4" src="img/natlib-logo-orange.png" alt="" width="72" height="57">
+        <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
+
+        <div class="form-floating">
+          <input type="text" class="form-control" id="username" placeholder="Username">
+          <label for="floatingInput">Username</label>
+        </div>
+        <div class="form-floating">
+          <input type="password" class="form-control" id="password" placeholder="Password">
+          <label for="floatingPassword">Password</label>
+        </div>
+
+        <!-- <div class="checkbox mb-3">
+          <label>
+            <input type="checkbox" value="remember-me" id="remember-credential"> Remember me
+          </label>
+        </div> -->
+        <button class="w-100 btn btn-lg btn-primary" onclick="login();">Sign in</button>
+        <p class="mt-5 mb-3 text-muted">&copy; 2021-2025 DIA.GOVT.NZ</p>
+      </div>
+    </main>
+
+    <script src="./dist/jquery-3.5.1.min.js"></script>
+    <script src="./dist/bootstrap.bundle.js"></script>
+    <script src="./dist/toastr.min.js"></script>
+    <script src="./app/dashboard.js"></script>
+
+    <script type="text/javascript">
+      function login(){
+        var cmd={};
+        cmd.username=$('#username').val();
+        cmd.password=$('#password').val();
+
+        if (!cmd.username || !cmd.password) {
+          toastr.error('Please input the username and password');
+          return;
+        }
+
+        fetch('auth/login.json', {
+          method: 'POST',
+          credentials: 'include',
+          redirect: 'follow',
+          headers: {'Content-Type': 'application/json'},
+          body: JSON.stringify(cmd)
+        }).then((response) => {
+          return response.json();
+        }).then((response) => {
+          if (response.rspCode==0) {
+              if (parent && parent.isParentAvailable && parent.isParentAvailable()) {
+                parent.loginCallback();
+              }else{
+                window.location.replace('home.html');
+              }
+          }else{
+            toastr.error(response.rspMsg);
+          }
+          
+        }).catch((e) => {
+          toastr.error(e);
+        });
+      }
+
+      $(function(){
+          toastr.options=Object.assign(toastr.options,{
+            "positionClass": "toast-top-center",
+            "closeButton": true,
+            "showDuration": "300",
+            "hideDuration": "1000",
+            "timeOut": "5000",
+            "extendedTimeOut": "1000",
+          });
+      });
+      
+    </script>
+  </body>
+</html>
